+++ lua/barbar/constants.lua
@@ -1,0 +1,18 @@
+
+--- @class barbar.constants.animation
+--- @field OPEN_DELAY integer
+--- @field OPEN_DURATION integer
+--- @field CLOSE_DURATION integer
+--- @field SCROLL_DURATION integer
+local ANIMATION = {
+  OPEN_DELAY = 10,
+  OPEN_DURATION = 150,
+  CLOSE_DURATION = 150,
+  SCROLL_DURATION = 200,
+}
+
+local constants = {
+  ANIMATION = ANIMATION,
+}
+
+return constants
